# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "evidence-validation API",
    "version": "0.1.0",
    "description": "Unfunded drafts and pilot reviews. No payment or distributed publication endpoints."
  },
  "servers": [
    {
      "url": "https://evidence-validation.astra-via.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped evk_ credential or existing Astra Supabase access token"
      }
    },
    "schemas": {
      "CreateJob": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "claim",
          "quote",
          "sourceUrl",
          "location",
          "relationship",
          "rewardCents"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 180
          },
          "claim": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10000
          },
          "quote": {
            "type": "string",
            "minLength": 5,
            "maxLength": 10000
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri",
            "description": "HTTP/HTTPS reference; not fetched"
          },
          "location": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "relationship": {
            "type": "string",
            "enum": [
              "supports",
              "challenges",
              "context"
            ]
          },
          "rewardCents": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1000000
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "public"
            ],
            "default": "private"
          }
        }
      },
      "SubmitReview": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "verdict",
          "reasoning",
          "quote",
          "location",
          "conflictFree"
        ],
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "supports",
              "challenges",
              "context",
              "insufficient",
              "mismatch"
            ]
          },
          "reasoning": {
            "type": "string",
            "minLength": 40,
            "maxLength": 10000
          },
          "quote": {
            "type": "string",
            "minLength": 5,
            "maxLength": 10000
          },
          "location": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "conflictFree": {
            "type": "boolean",
            "const": true
          }
        }
      }
    }
  },
  "paths": {
    "/api/v1/jobs": {
      "get": {
        "operationId": "listJobs",
        "description": "Requires jobs:read. Unfunded pilot operation.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "404": {
            "description": "Record inaccessible or absent"
          }
        }
      },
      "post": {
        "operationId": "createJob",
        "description": "Requires jobs:write. Unfunded pilot operation.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "404": {
            "description": "Record inaccessible or absent"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJob"
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{id}/reviews": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "operationId": "submitReview",
        "description": "Requires reviews:write. Unfunded pilot operation.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "404": {
            "description": "Record inaccessible or absent"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReview"
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{id}/bundle": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "exportBundle",
        "description": "Requires bundles:read. Unfunded pilot operation.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "404": {
            "description": "Record inaccessible or absent"
          }
        }
      }
    }
  }
}